+++ node/_util/async.ts
@@ -1,4 +1,6 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
+// This module is vendored from std/async/deferred.ts and std/async/delay.ts
+// (with some modifications)
 
 export interface Deferred<T> extends Promise<T> {
   readonly state: "pending" | "fulfilled" | "rejected";
@@ -27,4 +29,27 @@ export function deferred<T>(): Deferred<T> {
   });
   Object.defineProperty(promise, "state", { get: () => state });
   return Object.assign(promise, methods) as Deferred<T>;
+}
+
+/** Resolve a Promise after a given amount of milliseconds. */
+export function delay(
+  ms: number,
+  options: { signal?: AbortSignal } = {},
+): Promise<void> {
+  const { signal } = options;
+  if (signal?.aborted) {
+    return Promise.reject(new DOMException("Delay was aborted.", "AbortError"));
+  }
+  return new Promise((resolve, reject) => {
+    const abort = () => {
+      clearTimeout(i);
+      reject(new DOMException("Delay was aborted.", "AbortError"));
+    };
+    const done = () => {
+      signal?.removeEventListener("abort", abort);
+      resolve();
+    };
+    const i = setTimeout(done, ms);
+    signal?.addEventListener("abort", abort, { once: true });
+  });
 }
